Fix synthetic generation
